Add files via upload
A short python script that was planed to calculate lottery numbers. Sadly my hardware isn´t powerful enough to calculate all numbers up to 999,999 so I shortened it to just 999. Now it tries to hit every number in the range from 0 to 999 with randomized numbers and safes how many numbers it had to generate to get the specific number. For every number it does a hundred run throughs and safes every result. Then it calculates the average tries from those hundred run throughs and after it did this for all values my program prints out the number with the lowest average tries.